+++ mmdeploy/codebase/mmcls/models/backbones/__init__.py
@@ -1,13 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .shufflenet_v2 import shufflenetv2_backbone__forward__default
-from .swin_transformer import (shift_window_msa__forward__default,
-                               shift_window_msa__get_attn_mask__default,
-                               window_msa__forward__default)
 from .vision_transformer import visiontransformer__forward__ncnn
 
 __all__ = [
     'shufflenetv2_backbone__forward__default',
-    'visiontransformer__forward__ncnn',
-    'shift_window_msa__get_attn_mask__default',
-    'shift_window_msa__forward__default', 'window_msa__forward__default'
+    'visiontransformer__forward__ncnn'
 ]
