+++ mmdeploy/codebase/mmcls/models/utils/__init__.py
@@ -1,4 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .attention import multiheadattention__forward__ncnn
+from .attention import (multiheadattention__forward__ncnn,
+                        shift_window_msa__forward__default,
+                        shift_window_msa__get_attn_mask__default,
+                        window_msa__forward__default)
 
-__all__ = ['multiheadattention__forward__ncnn']
+__all__ = [
+    'multiheadattention__forward__ncnn',
+    'shift_window_msa__get_attn_mask__default',
+    'shift_window_msa__forward__default', 'window_msa__forward__default'
+]
